requirements.txt
pykinect-azure
opencv-python
numpy
pandas
rich
tqdm
pyinstaller
imagezmq
sounddevice